# auth-service/internal/http/auth/login.go
package auth

import (
	"encoding/json"
	"net/http"

	authdomain "github.com/PetoAdam/homenavi/auth-service/internal/auth"
	"github.com/PetoAdam/homenavi/auth-service/internal/errors"
	authtransport "github.com/PetoAdam/homenavi/auth-service/internal/http/auth/transport"
	sharedtransport "github.com/PetoAdam/homenavi/auth-service/internal/http/transport"
	clientsinfra "github.com/PetoAdam/homenavi/auth-service/internal/infra/clients"
)

type LoginHandler struct {
	authService  *authdomain.Service
	userService  *clientsinfra.UserClient
	emailService *clientsinfra.EmailClient
}

func NewLoginHandler(authService *authdomain.Service, userService *clientsinfra.UserClient, emailService *clientsinfra.EmailClient) *LoginHandler {
	return &LoginHandler{
		authService:  authService,
		userService:  userService,
		emailService: emailService,
	}
}

func (h *LoginHandler) HandleLoginStart(w http.ResponseWriter, r *http.Request) {
	var req authtransport.LoginStartRequest
	if err := sharedtransport.ParseAndValidateJSON(r, &req); err != nil {
		errors.WriteError(w, errors.BadRequest(err.Error()))
		return
	}

	result, err := h.authService.StartLogin(req.Email, req.Password, h.userService, h.emailService)
	if err != nil {
		if appErr, ok := err.(*errors.AppError); ok {
			errors.WriteError(w, appErr)
			return
		}
		errors.WriteError(w, errors.InternalServerError("failed to start login", err))
		return
	}

	response := authtransport.LoginResponse{
		TwoFARequired: result.TwoFARequired,
		UserID:        result.UserID,
		TwoFAType:     result.TwoFAType,
	}
	if !result.TwoFARequired {
		SetSessionCookies(w, r, &authdomain.TokenPair{AccessToken: result.AccessToken, RefreshToken: result.RefreshToken})
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
}

func (h *LoginHandler) HandleLoginFinish(w http.ResponseWriter, r *http.Request) {
	var req authtransport.LoginFinishRequest
	if err := sharedtransport.ParseAndValidateJSON(r, &req); err != nil {
		errors.WriteError(w, errors.BadRequest(err.Error()))
		return
	}

	var tokens *authdomain.TokenPair
	var err error
	if req.RecoveryCode {
		tokens, err = h.authService.FinishLoginWithRecoveryCode(req.UserID, req.Code, h.userService, h.userService)
	} else {
		tokens, err = h.authService.FinishLogin(req.UserID, req.Code, h.userService)
	}
	if err != nil {
		if appErr, ok := err.(*errors.AppError); ok {
			errors.WriteError(w, appErr)
			return
		}
		errors.WriteError(w, errors.InternalServerError("failed to complete login", err))
		return
	}

	SetSessionCookies(w, r, tokens)
	response := authtransport.LoginResponse{}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
}