# doc/openapi.yaml
openapi: 3.1.0
info:
  title: Homenavi External API
  version: 0.5.16
  description: |
    HTTP API exposed through the Homenavi API gateway. This document covers the
    supported public control-plane contract. Event streams are described in
    doc/asyncapi.yaml.
servers:
  - url: /
    description: Current deployment ingress origin
  - url: http://localhost
    description: Local Compose ingress
security:
  - bearerAuth: []
...
paths:
...
  /api/auth/login/start:
    post:
      tags: [Authentication]
      security: []
      operationId: startLogin
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginStartRequest' }
      responses:
        '200': { $ref: '#/components/responses/JSONSuccess' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/auth/login/finish:
    post:
      tags: [Authentication]
      security: []
      operationId: finishLogin
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginFinishRequest' }
      responses:
        '200': { $ref: '#/components/responses/JSONSuccess' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT sent as an Authorization bearer token. Browser clients may also use the auth cookie.
...
  responses:
    JSONSuccess:
      description: Successful JSON response
      content: { application/json: { schema: { type: object, additionalProperties: true } } }
...
    BadRequest: { description: Request validation failed }
    Unauthorized: { description: Authentication is required or invalid }
...
  schemas:
...
    LoginStartRequest:
      type: object
      required: [email]
      properties: { email: { type: string, format: email } }
    LoginFinishRequest:
      type: object
      required: [email, code]
      properties:
        email: { type: string, format: email }
        code: { type: string }
...